Implement hex to_decimal. Refactor NumberConverterBaseClass and subclasses, as int is too small for deadbeef

+++ src/main/java/com/example/application/components/numberConverter/NumberConverterInterface.java
@@ -1,6 +1,6 @@
 package com.example.application.components.numberConverter;
 
 public abstract interface NumberConverterInterface {
-    public Integer to_decimal(String value);
-    public String from_decimal(Integer value);
+    public Long to_decimal(String value);
+    public String from_decimal(Long value);
 }
